analyzer-service/requirements.txt
grpcio
grpcio-tools
protobuf
PyMuPDF
language_tool_python
